feat: fix seo metadata on map page
