+++ home/models.py
@@ -1,9 +1,10 @@
 from wagtail.admin.edit_handlers import FieldPanel
 from wagtail.images.edit_handlers import ImageChooserPanel
 from logbooks.views import pages_for_tag
-from logbooks.models.mixins import ArticleSeoMixin, SeoMetadataMixin
+from logbooks.models.mixins import ArticleSeoMixin
 from logbooks.models.tag_cloud import TagCloud
 from smartforests.models import CmsImage
+from smartforests.mixins import SeoMetadataMixin
 from django.db import models
 from wagtail.core.fields import RichTextField
 from wagtail.core.models import Page
